added check to ensure variable is empty or not

+++ amps/ags/ci/scripts/cleanup.sh
@@ -3,20 +3,22 @@
 echo "=========================== Starting Cleanup Script ==========================="
 PS4="\[\e[35m\]+ \[\e[m\]"
 set -vx
+set -e
 pushd "$(dirname "${BASH_SOURCE[0]}")/../"
 
+if [[ -z "$S3_BUCKET_NAME" || -z "$S3_BUCKET2_NAME" ]]; then
+    echo "Error: S3_BUCKET_NAME or S3_BUCKET2_NAME is not set."
+    exit 1
+fi
 
 # Stop and remove the containers
 docker ps -a -q | xargs -l -r docker stop
 docker ps -a -q | xargs -l -r docker rm
 
-pip3 install awscli
-printf "${AGS_AWS_ACCESS_KEY_ID}\n${AGS_AWS_SECRET_ACCESS_KEY}\n\n\n" | aws configure
-
 if aws s3 ls "s3://${S3_BUCKET_NAME}" &>/dev/null; then
     echo "Bucket '${S3_BUCKET_NAME}' exists. Deleting..."
-    aws s3 rb "s3://${S3_BUCKET_NAME}" --force
-    aws s3 rb "s3://${S3_BUCKET2_NAME}" --force
+    aws s3 rb "s3://${S3_BUCKET_NAME}" --force || true
+    aws s3 rb "s3://${S3_BUCKET2_NAME}" --force || true
     echo "Bucket deleted."
 else
     echo "Bucket '${S3_BUCKET_NAME}' does not exist."
